refactor(spec): split validate-data into evidence + checks modules
Extract FieldEvidence construction (resolveFieldEvidence, effectiveChannel)
and the tier-2 dataChecks walk (STAT_COLUMNS) from the 531-line validate-data
barrel so evidence can be shared with lint without dragging the check walk.
Public package exports stay on validate-data.js. Characterization covers
maxBytes short-circuit and evidence-failure ordering with structural stages.
