+++ packages/spec/src/validate-data-checks.ts
@@ -1,0 +1,242 @@
+/**
+ * Tier-2 data-aware walk — field existence, all-null columns, {stat} columns,
+ * geom/stat type rules, and scale/type compatibility. Consumes a FieldEvidenceMap
+ * built by validate-data-evidence.ts (or builds one when not pre-resolved).
+ */
+import type { SpecError } from "./errors.js";
+import { didYouMean } from "./errors.js";
+import type { Aes, ChannelName, ColorScaleSpec, PositionScaleSpec } from "./schema.js";
+import { CHANNELS, GEOM_DEFAULTS } from "./schema.js";
+import type { ProfileFieldType, ValidateLimits, ValidateOptions } from "./validate-data.js";
+import {
+  effectiveChannel,
+  resolveFieldEvidence,
+  type ResolveFieldEvidenceResult,
+} from "./validate-data-evidence.js";
+
+/** Columns generated by each stat ({ stat } channel resolution contract).
+ *  Documented per stat in its @ggsvelte/core module header. */
+export const STAT_COLUMNS: Record<string, readonly string[]> = {
+  identity: [],
+  count: ["count"],
+  bin: ["count", "density", "ncount", "ndensity"],
+  density: ["density", "count", "scaled", "ndensity"],
+  smooth: ["y", "ymin", "ymax", "se"],
+  boxplot: ["ymin", "lower", "middle", "upper", "ymax"],
+  summary: ["y", "ymin", "ymax"],
+};
+
+function isRecord(v: unknown): v is Record<string, unknown> {
+  return typeof v === "object" && v !== null && !Array.isArray(v);
+}
+
+const AXIS_CHANNELS = ["x", "y"] as const;
+const COLOR_CHANNELS = ["color", "fill"] as const;
+
+export function dataChecks(
+  spec: Record<string, unknown>,
+  options: ValidateOptions,
+  limits: ValidateLimits,
+  /** Pre-resolved evidence from validate() so lint can share the same pass. */
+  preResolved?: ResolveFieldEvidenceResult,
+): SpecError[] {
+  const errors: SpecError[] = [];
+
+  const resolved = preResolved ?? resolveFieldEvidence(spec, options, limits);
+  if (resolved.status === "errors") return resolved.errors;
+  if (resolved.status === "none") return errors; // no data, no profile: nothing to check
+  const fields = resolved.fields;
+
+  const available = [...fields.keys()];
+  const plotAes = isRecord(spec["aes"]) ? (spec["aes"] as Aes) : undefined;
+  const layers = Array.isArray(spec["layers"]) ? (spec["layers"] as unknown[]) : [];
+
+  interface AxisUse {
+    field: string;
+    path: string;
+  }
+  const axisFields: Record<"x" | "y", AxisUse[]> = { x: [], y: [] };
+  const colorFields: Record<"color" | "fill", AxisUse[]> = { color: [], fill: [] };
+
+  for (let i = 0; i < layers.length; i++) {
+    const layer = layers[i];
+    if (!isRecord(layer)) continue;
+    const geom = typeof layer["geom"] === "string" ? layer["geom"] : "";
+    const layerAes = isRecord(layer["aes"]) ? (layer["aes"] as Aes) : undefined;
+    const defaultStat = GEOM_DEFAULTS[geom as keyof typeof GEOM_DEFAULTS]?.stat ?? "identity";
+    const stat = typeof layer["stat"] === "string" ? layer["stat"] : defaultStat;
+
+    // --- geom/stat field-type rules (M2 statistical layer) -----------------
+    const fieldTypeOf = (channel: ChannelName): [string, ProfileFieldType] | null => {
+      const m = effectiveChannel(plotAes, layerAes, channel);
+      if (m === undefined || !("field" in m)) return null;
+      const t = fields.get(m.field)?.type;
+      return t === undefined || t === null ? null : [m.field, t];
+    };
+    const typeError = (channel: ChannelName, message: string, fixDesc: string) => {
+      errors.push({
+        code: "channel-type-mismatch",
+        path: `/layers/${i}/aes/${channel}`,
+        message,
+        fix: { description: fixDesc },
+      });
+    };
+    if (geom === "smooth") {
+      for (const channel of ["x", "y"] as const) {
+        const info = fieldTypeOf(channel);
+        if (info !== null && (info[1] === "nominal" || info[1] === "ordinal")) {
+          typeError(
+            channel,
+            `The smooth stat needs quantitative x and y, but field "${info[0]}" (${channel}) is ${info[1]}.`,
+            "Map the channel to a numeric field, or use a boxplot for discrete x.",
+          );
+        }
+      }
+    }
+    if (geom === "boxplot") {
+      const x = fieldTypeOf("x");
+      if (x !== null && (x[1] === "quantitative" || x[1] === "temporal")) {
+        typeError(
+          "x",
+          `The boxplot geom needs a DISCRETE x this milestone, but field "${x[0]}" is ${x[1]}.`,
+          "Map x to a categorical field (strings), or bin the values into labeled categories first.",
+        );
+      }
+      const y = fieldTypeOf("y");
+      if (y !== null && (y[1] === "nominal" || y[1] === "ordinal")) {
+        typeError(
+          "y",
+          `The boxplot stat needs a quantitative y, but field "${y[0]}" is ${y[1]}.`,
+          "Map y to a numeric field.",
+        );
+      }
+    }
+    if (geom === "histogram" || geom === "density" || (geom === "bar" && stat === "bin")) {
+      const x = fieldTypeOf("x");
+      if (x !== null && (x[1] === "nominal" || x[1] === "ordinal")) {
+        typeError(
+          "x",
+          `The ${geom === "density" ? "density" : "bin"} stat needs a continuous x, but field "${x[0]}" is ${x[1]}.`,
+          'Use geom "bar" (the count stat) to count categories instead.',
+        );
+      }
+    }
+    if (geom === "errorbar") {
+      for (const channel of ["ymin", "ymax"] as const) {
+        const info = fieldTypeOf(channel);
+        if (info !== null && (info[1] === "nominal" || info[1] === "ordinal")) {
+          typeError(
+            channel,
+            `The errorbar geom needs quantitative bounds, but field "${info[0]}" (${channel}) is ${info[1]}.`,
+            "Map the channel to a numeric field.",
+          );
+        }
+      }
+    }
+
+    for (const channel of CHANNELS) {
+      const mapped = effectiveChannel(plotAes, layerAes, channel);
+      if (mapped === undefined) continue;
+      const path = `/layers/${i}/aes/${channel}`;
+
+      if ("stat" in mapped) {
+        const generated = STAT_COLUMNS[stat] ?? [];
+        if (!generated.includes(mapped.stat)) {
+          errors.push({
+            code: "unknown-stat-column",
+            path,
+            message:
+              generated.length === 0
+                ? `Channel "${channel}" maps stat column "${mapped.stat}", but this layer's stat ("${stat}") generates no columns.`
+                : `Channel "${channel}" maps stat column "${mapped.stat}", but this layer's stat ("${stat}") generates: ${generated.join(", ")}.`,
+            ...(generated.length > 0 && { allowed: [...generated] }),
+          });
+        }
+        continue;
+      }
+      if (!("field" in mapped)) continue;
+
+      const info = fields.get(mapped.field);
+      if (info === undefined) {
+        const suggestion = didYouMean(mapped.field, available);
+        errors.push({
+          code: "unknown-field",
+          path,
+          message:
+            `Unknown field "${mapped.field}" (available: ${available.join(", ") || "none"}).` +
+            (suggestion === undefined ? "" : ` Did you mean "${suggestion}"?`),
+          allowed: available,
+          ...(suggestion !== undefined && {
+            fix: {
+              description: `Map "${channel}" to "${suggestion}".`,
+              example: { field: suggestion },
+            },
+          }),
+        });
+        continue;
+      }
+      if (info.allNull) {
+        errors.push({
+          code: "all-null-column",
+          path,
+          message: `Field "${mapped.field}" contains only null values; the "${channel}" channel cannot be drawn from it.`,
+        });
+        continue;
+      }
+      if (channel === "x" || channel === "y")
+        axisFields[channel].push({ field: mapped.field, path });
+      if (channel === "color" || channel === "fill") {
+        colorFields[channel].push({ field: mapped.field, path });
+      }
+    }
+  }
+
+  // --- scale/type compatibility ----------------------------------------------
+  const scales = isRecord(spec["scales"]) ? spec["scales"] : undefined;
+  const typeOf = (field: string) => fields.get(field)?.type ?? null;
+
+  for (const axis of AXIS_CHANNELS) {
+    const config = scales?.[axis] as PositionScaleSpec | undefined;
+    const declared = config?.type;
+    if (declared === undefined || declared === "band") continue;
+    for (const use of axisFields[axis]) {
+      const type = typeOf(use.field);
+      if (type === null) continue;
+      if (declared === "time" && type !== "temporal") {
+        errors.push({
+          code: "scale-type-mismatch",
+          path: use.path,
+          message: `scales.${axis}.type is "time" but field "${use.field}" is ${type}; time scales need temporal values (Dates or ISO 8601 strings).`,
+        });
+      } else if (
+        (declared === "log" || declared === "linear") &&
+        (type === "nominal" || type === "ordinal")
+      ) {
+        errors.push({
+          code: "scale-type-mismatch",
+          path: use.path,
+          message: `scales.${axis}.type is "${declared}" but field "${use.field}" is ${type}; use a band scale (or a quantitative field).`,
+          fix: { description: `Set scales.${axis}.type to "band".` },
+        });
+      }
+    }
+  }
+
+  for (const channel of COLOR_CHANNELS) {
+    const config = scales?.[channel] as ColorScaleSpec | undefined;
+    if (config?.type !== "sequential") continue;
+    for (const use of colorFields[channel]) {
+      const type = typeOf(use.field);
+      if (type === "nominal" || type === "ordinal") {
+        errors.push({
+          code: "scale-type-mismatch",
+          path: use.path,
+          message: `scales.${channel}.type is "sequential" but field "${use.field}" is ${type}; sequential color ramps need quantitative values.`,
+          fix: { description: `Set scales.${channel}.type to "ordinal".` },
+        });
+      }
+    }
+  }
+
+  return errors;
+}
