+++ packages/spec/src/validate.ts
@@ -6,7 +6,8 @@
  * the same schemas that emit `schema/v0.json` — one artifact, no drift.
  * Raw TypeBox union noise is mapped to the agent error contract in
  * validate-map-errors.ts. Data-free grammar rules live in
- * validate-structure.ts. Data-aware checks live in validate-data.ts.
+ * validate-structure.ts. Data-aware checks live in validate-data*.ts
+ * (evidence + checks modules, barrel at validate-data.ts).
  *
  * Output: `{ ok: true, spec }` or `{ ok: false, errors: SpecError[] }` with
  * the agent error contract from errors.ts. Messages are snapshot-tested.
